+++ src/components/Editor/Runners/RunnerFactory.jsx
@@ -3,7 +3,11 @@ import PyodideRunner from "./PyodideRunner/PyodideRunner";
 import PythonRunner from "./PythonRunner/PythonRunner";
 import HtmlRunner from "./HtmlRunner/HtmlRunner";
 
-const RunnerFactory = ({ projectType, usePyodide }) => {
+const RunnerFactory = ({
+  projectType,
+  usePyodide,
+  outputPanels = ["text", "visual"],
+}) => {
   const Runner = () => {
     if (projectType === "html") {
       return HtmlRunner;
@@ -16,7 +20,9 @@ const RunnerFactory = ({ projectType, usePyodide }) => {
 
   const Selected = Runner();
 
-  return <Selected />;
+  const props = projectType === "html" ? {} : { outputPanels };
+
+  return <Selected {...props} />;
 };
 
 export default RunnerFactory;