+++ src/redux/EditorSlice.js
@@ -12,19 +12,24 @@ import {
   createRemix,
   deleteProject,
   readProjectList,
+  loadAssets,
 } from "../utils/apiCallHandler";
 
 export const syncProject = (actionName) =>
   createAsyncThunk(
     `editor/${actionName}Project`,
     async (
-      { project, identifier, locale, accessToken, autosave },
+      { project, identifier, locale, accessToken, autosave, assetsOnly },
       { rejectWithValue },
     ) => {
       let response;
       switch (actionName) {
         case "load":
-          response = await readProject(identifier, locale, accessToken);
+          if (assetsOnly) {
+            response = await loadAssets(identifier, locale, accessToken);
+          } else {
+            response = await readProject(identifier, locale, accessToken);
+          }
           break;
         case "loadRemix":
           response = await loadRemix(identifier, accessToken);
@@ -97,6 +102,7 @@ export const EditorSlice = createSlice({
     codeHasBeenRun: false,
     drawTriggered: false,
     isEmbedded: false,
+    isOutputOnly: false,
     browserPreview: false,
     isSplitView: true,
     isThemeable: true,
@@ -125,6 +131,7 @@ export const EditorSlice = createSlice({
     deleteProjectModalShowing: false,
     sidebarShowing: true,
     modals: {},
+    errorDetails: {},
   },
   reducers: {
     closeFile: (state, action) => {
@@ -187,6 +194,9 @@ export const EditorSlice = createSlice({
     },
     setEmbedded: (state, _action) => {
       state.isEmbedded = true;
+    },
+    setIsOutputOnly: (state, action) => {
+      state.isOutputOnly = action.payload;
     },
     setBrowserPreview: (state, _action) => {
       state.browserPreview = true;
@@ -367,6 +377,9 @@ export const EditorSlice = createSlice({
     disableTheming: (state) => {
       state.isThemeable = false;
     },
+    setErrorDetails: (state, action) => {
+      state.errorDetails = action.payload;
+    },
   },
   extraReducers: (builder) => {
     builder.addCase("editor/saveProject/pending", (state) => {
@@ -456,6 +469,7 @@ export const {
   setFocussedFileIndex,
   setPage,
   setEmbedded,
+  setIsOutputOnly,
   setBrowserPreview,
   setError,
   setIsSplitView,
@@ -498,6 +512,7 @@ export const {
   showSidebar,
   hideSidebar,
   disableTheming,
+  setErrorDetails,
 } = EditorSlice.actions;
 
 export default EditorSlice.reducer;