+++ src/redux/EditorSlice.test.js
@@ -2,6 +2,8 @@ import {
   createOrUpdateProject,
   createRemix,
   deleteProject,
+  loadAssets,
+  readProject,
   readProjectList,
 } from "../utils/apiCallHandler";
 
@@ -16,6 +18,8 @@ import reducer, {
   updateComponentName,
   loadProjectList,
   setLoadRemixDisabled,
+  setIsOutputOnly,
+  setErrorDetails,
 } from "./EditorSlice";
 
 jest.mock("../utils/apiCallHandler");
@@ -54,6 +58,41 @@ test("Action setLoadRemixDisabled sets loadRemixDisabled to false", () => {
   expect(reducer(previousState, setLoadRemixDisabled(false))).toEqual(
     expectedState,
   );
+});
+
+test("Action setIsOutputOnly sets isOutputOnly to true", () => {
+  const previousState = {
+    isOutputOnly: false,
+  };
+  const expectedState = {
+    isOutputOnly: true,
+  };
+  expect(reducer(previousState, setIsOutputOnly(true))).toEqual(expectedState);
+});
+
+test("Action setOutputOnly sets isOutputOnly to false", () => {
+  const previousState = {
+    isOutputOnly: true,
+  };
+  const expectedState = {
+    isOutputOnly: false,
+  };
+  expect(reducer(previousState, setIsOutputOnly(false))).toEqual(expectedState);
+});
+
+test("Action setErrorDetails sets errorDetails to true", () => {
+  const previousState = {
+    errorDetails: {},
+  };
+  const expectedState = {
+    errorDetails: { type: "Interrupted", message: "Some error message" },
+  };
+  expect(
+    reducer(
+      previousState,
+      setErrorDetails({ type: "Interrupted", message: "Some error message" }),
+    ),
+  ).toEqual(expectedState);
 });
 
 test("Showing rename modal sets file state and showing status", () => {
@@ -612,5 +651,56 @@ describe("Updating file name", () => {
         updateComponentName({ key: 1, name: "my_file", extension: "py" }),
       ),
     ).toEqual(expectedState);
+  });
+});
+
+describe("Loading a project", () => {
+  const dispatch = jest.fn();
+  const identifier = "my-project-identifier";
+  const accessToken = "myToken";
+  const locale = "es-LA";
+
+  let initialState;
+  let loadThunk;
+
+  beforeEach(() => {
+    initialState = { editor: {}, auth: {} };
+    loadThunk = syncProject("load");
+  });
+
+  describe("when assetsOnly is false", () => {
+    let loadAction;
+
+    beforeEach(() => {
+      loadAction = loadThunk({
+        identifier,
+        locale,
+        accessToken,
+        assetsOnly: false,
+      });
+    });
+
+    test("readProject is called", async () => {
+      await loadAction(dispatch, () => initialState);
+      expect(readProject).toHaveBeenCalledWith(identifier, locale, accessToken);
+    });
+  });
+
+  describe("when assetsOnly is true", () => {
+    let loadAction;
+
+    beforeEach(() => {
+      loadAction = loadThunk({
+        identifier,
+        locale,
+        accessToken,
+        assetsOnly: true,
+      });
+    });
+
+    test("loadAssets is called", async () => {
+      await loadAction(dispatch, () => initialState);
+      expect(loadAssets).toHaveBeenCalledWith(identifier, locale, accessToken);
+    });
   });
 });